start reports and data prep workers
