+++ es/index.html
@@ -30,9 +30,9 @@
   <script type="application/javascript" src="/static/js/locales/es.js" async></script>
   <script type="module" defer src="/static/js/plots.js"></script>
   <script type="module" defer src="/static/js/ui.js"></script>
-  <script type="module" defer src="/static/js/cache.js"></script>
-  <script type="module" defer src="/static/js/rfsZarrFetcher.js"></script>
-  <script type="module" defer src="/static/js/data.js"></script>
+  <script type="module" defer src="/js/data/cache.js"></script>
+  <script type="module" defer src="/js/data/rfsZarrFetcher.js"></script>
+  <script type="module" defer src="/js/data/data.js"></script>
   <script type="module" defer src="/static/js/map.js"></script>
   <script type="module" defer src="/static/js/main.js"></script>
 </head>
