+++ static/js/data/rfsZarrFetcher.js
