+++ static/js/main.js
@@ -1,7 +1,8 @@
-import {displayLoadingStatus, displayRiverNumber, lang, riverIdInput} from "./ui.js";
-import {fetchData, updateDownloadLinks} from "./data.js";
+import {clearCharts, displayLoadingStatus, displayRiverNumber, divModalCharts, inputForecastDate, lang, riverIdInput, updateDownloadLinks} from "./ui.js";
+import {getAndCacheForecast, getAndCacheRetrospective, getAndCacheReturnPeriods} from "./data/data.js";
 import {bookmarks} from "./bookmarks.js";
-import {LoadStatus, RiverId, UseBiasCorrected, UseSimpleForecast, UseShowExtraRetroGraphs} from "./states/state.js";
+import {LoadStatus, RiverId, UseBiasCorrected, UseShowExtraRetroGraphs, UseSimpleForecast} from "./states/state.js";
+import {plotAllForecast, plotAllRetro} from "./plots.js";
 
 //////////////////////////////////////////////////////////////////////// INITIAL LOAD
 M.AutoInit();
@@ -12,6 +13,41 @@ M.Dropdown.init(document.querySelectorAll('.dropdown-trigger'), {
 });
 Plotly.setPlotConfig({'locale': lang})
 if (window.innerWidth < 800) M.toast({html: text.prompts.mobile, classes: "blue custom-toast-placement", displayLength: 7500})
+
+const fetchData = ({riverId, display = true} = {}) => {
+  if (display) M.Modal.getInstance(divModalCharts).open()
+  riverId = riverId || RiverId.get()
+  if (!riverId) return
+  const date = inputForecastDate.value.replaceAll("-", "")
+  const corrected = UseBiasCorrected.get()
+  const stats = UseSimpleForecast.get()
+  clearCharts('forecast')
+  clearCharts('retro')
+  LoadStatus.update({forecast: "load", retro: "load"})
+  Promise
+    .all([getAndCacheForecast({riverId, date, corrected}), getAndCacheReturnPeriods(riverId)])
+    .then(responses => {
+      plotAllForecast({forecast: responses[0], rp: responses[1], riverId, corrected, showStats: stats})
+      LoadStatus.update({forecast: "ready"})
+      updateDownloadLinks({forecast: responses[0], riverId})
+    })
+    .catch(error => {
+      console.error(error)
+      LoadStatus.update({forecast: "fail"})
+      clearCharts('forecast')
+    })
+  getAndCacheRetrospective({riverId, corrected})
+    .then(response => {
+      plotAllRetro({retro: response, riverId})
+      LoadStatus.update({retro: "ready"})
+      updateDownloadLinks({retro: response, riverId})
+    })
+    .catch(error => {
+      console.error(error)
+      LoadStatus.update({retro: "fail"})
+      clearCharts('retro')
+    })
+}
 
 // subscribers to RiverId changes - don't change the order
 RiverId.addSubscriber(LoadStatus.reset)
